requirements.txt
altair==5.5.0
attrs==25.1.0
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
einops==0.8.0
filelock==3.17.0
flatbuffers==25.1.24
fsspec==2025.2.0
gitdb==4.0.12
GitPython==3.1.44
huggingface-hub~=0.29.1
humanfriendly==10.0
idna==3.10
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
moondream==0.0.6
mpmath==1.3.0
narwhals==1.26.0
networkx==3.3
numpy==2.2.2
onnx==1.17.0
onnxruntime==1.20.1
packaging==24.2
pandas==2.2.3
pillow==10.4.0
protobuf==5.29.3
pyarrow==19.0.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2025.1
pyvips==2.2.3
pyvips-binary==8.16.0
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
safetensors==0.5.2
six==1.17.0
smmap==5.0.2
streamlit==1.42.0
sympy==1.13.1
tenacity==9.0.0
tokenizers==0.20.3
toml==0.10.2
torch==2.6.0+cpu
torchaudio==2.6.0+cpu
torchvision==0.21.0+cpu
tornado==6.4.2
tqdm==4.67.1
transformers==4.48.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
watchdog==6.0.0
altair==5.5.0
attrs==25.1.0
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
einops==0.8.0
filelock==3.17.0
flatbuffers==25.1.24
fsspec==2025.2.0
gitdb==4.0.12
GitPython==3.1.44
huggingface-hub==0.28.1
humanfriendly==10.0
idna==3.10
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
moondream==0.0.6
mpmath==1.3.0
narwhals==1.26.0
networkx==3.3
numpy==2.2.2
onnx==1.17.0
onnxruntime==1.20.1
packaging==24.2
pandas==2.2.3
pillow==10.4.0
protobuf==5.29.3
pyarrow==19.0.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
pyttsx3==2.98
pytz==2025.1
pyvips==2.2.3
pyvips-binary==8.16.0
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
safetensors==0.5.2
six==1.17.0
smmap==5.0.2
streamlit==1.42.0
streamlit-audiorec==0.1.3
sympy==1.13.1
tenacity==9.0.0
tokenizers==0.20.3
toml==0.10.2
torch
torchaudio
torchvision
tornado==6.4.2
tqdm==4.67.1
transformers==4.48.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
watchdog==6.0.0
altair==5.5.0
attrs==25.1.0
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
einops==0.8.0
filelock==3.17.0
flatbuffers==25.1.24
fsspec==2025.2.0
gitdb==4.0.12
GitPython==3.1.44
huggingface-hub==0.28.1
humanfriendly==10.0
idna==3.10
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
moondream==0.0.6
mpmath==1.3.0
narwhals==1.26.0
networkx==3.3
numpy==2.2.2
onnx==1.17.0
onnxruntime==1.20.1
packaging==24.2
pandas==2.2.3
pillow==10.4.0
protobuf==5.29.3
pyarrow==19.0.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
pyttsx3==2.98
pytz==2025.1
pyvips==2.2.3
pyvips-binary==8.16.0
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
safetensors==0.5.2
six==1.17.0
smmap==5.0.2
streamlit==1.42.0
sympy==1.13.1
tenacity==9.0.0
tokenizers==0.20.3
toml==0.10.2
torch==2.6.0+cpu
torchaudio==2.6.0+cpu
torchvision==0.21.0+cpu
tornado==6.4.2
tqdm==4.67.1
transformers==4.48.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
watchdog==6.0.0
altair==5.5.0
attrs==25.1.0
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
einops==0.8.0
filelock==3.17.0
flatbuffers==25.1.24
fsspec==2025.2.0
gitdb==4.0.12
GitPython==3.1.44
huggingface-hub==0.28.1
humanfriendly==10.0
idna==3.10
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
moondream==0.0.6
mpmath==1.3.0
narwhals==1.26.0
networkx==3.3
numpy==2.2.2
onnx==1.17.0
onnxruntime==1.20.1
packaging==24.2
pandas==2.2.3
pillow==10.4.0
protobuf==5.29.3
pyarrow==19.0.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
pyttsx3==2.98
pytz==2025.1
pyvips==2.2.3
pyvips-binary==8.16.0
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
safetensors==0.5.2
six==1.17.0
smmap==5.0.2
SpeechRecognition==3.14.1
streamlit==1.42.0
streamlit_mic_recorder==0.0.8
sympy==1.13.1
tenacity==9.0.0
tokenizers==0.20.3
toml==0.10.2
torch==2.6.0+cpu
torchaudio==2.6.0+cpu
torchvision==0.21.0+cpu
tornado==6.4.2
tqdm==4.67.1
transformers==4.48.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
watchdog==6.0.0
accelerate==1.4.0
altair==5.5.0
attrs==25.1.0
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
diskcache==5.6.3
einops==0.8.0
filelock==3.17.0
flatbuffers==25.1.24
fsspec==2025.2.0
gitdb==4.0.12
GitPython==3.1.44
huggingface-hub==0.29.1
humanfriendly==10.0
idna==3.10
importlib_metadata==8.6.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
llama_cpp_python==0.3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
moondream==0.0.6
mpmath==1.3.0
narwhals==1.26.0
networkx==3.3
numpy==2.2.2
onnx==1.17.0
onnxruntime==1.20.1
packaging==24.2
pandas==2.2.3
pillow==10.4.0
protobuf==5.29.3
psutil==7.0.0
pyarrow==19.0.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
pyttsx3==2.98
pytz==2025.1
pyvips==2.2.3
pyvips-binary==8.16.0
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
safetensors==0.5.2
six==1.17.0
smmap==5.0.2
SpeechRecognition==3.14.1
streamlit==1.42.0
streamlit_mic_recorder==0.0.8
sympy==1.13.1
tenacity==9.0.0
tokenizers==0.20.3
toml==0.10.2
torchaudio==2.6.0+cpu
torchvision==0.21.0+cpu
tornado==6.4.2
tqdm==4.67.1
transformers==4.48.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
watchdog==6.0.0
zipp==3.21.0
